With strict warnings on: Fixing Xcode warnings about implicit cast from `double` -> `float`, and loss of precision.

+++ MNColorKit/RGBX.h
@@ -10,9 +10,9 @@
 
 static __inline UIColor *RGBXA(NSUInteger rgb, CGFloat alpha)
 {
-    return [UIColor colorWithRed:((float)((rgb & 0xFF0000) >> 16))/255.0
-                           green:((float)((rgb & 0xFF00) >> 8))/255.0
-                            blue:((float)(rgb & 0xFF))/255.0
+    return [UIColor colorWithRed:(CGFloat)(((rgb & 0xFF0000) >> 16) / 255.0)
+                           green:(CGFloat)(((rgb & 0xFF00) >> 8) / 255.0)
+                            blue:(CGFloat)((rgb & 0xFF) / 255.0)
                            alpha:alpha];
 }
